still tracking down bug in the draw grid funciton after moving from engine to main activity due to need for access to imageView variable after creation.
